renaming according to naming conventions
