+++ wls-gui-wahllokalsystem/src/views/wahlvorbereitung/WahlumgebungView.vue
@@ -1,8 +1,8 @@
 <template>
-  <wahlumgebung-uwb-card
+  <the-wahlumgebung-uwb-card
     v-if="currentUserWahlbezirksArt == WahlbezirksArtEnum.UWB"
   />
-  <wahlumgebung-bwb-card
+  <the-wahlumgebung-bwb-card
     v-if="currentUserWahlbezirksArt == WahlbezirksArtEnum.BWB"
   />
 </template>
@@ -10,8 +10,8 @@
 <script setup lang="ts">
 import { storeToRefs } from "pinia";
 
-import WahlumgebungBwbCard from "@/components/wahlvorbereitung/WahlumgebungBwbCard.vue";
-import WahlumgebungUwbCard from "@/components/wahlvorbereitung/WahlumgebungUwbCard.vue";
+import TheWahlumgebungBwbCard from "@/components/wahlvorbereitung/TheWahlumgebungBwbCard.vue";
+import TheWahlumgebungUwbCard from "@/components/wahlvorbereitung/TheWahlumgebungUwbCard.vue";
 import { useUserStore } from "@/stores/userStore.ts";
 import { WahlbezirksArtEnum } from "@/types/wahlbezirksArtEnum.ts";
